custom-game-area/requirements.txt
Pillow==9.2.0